move `registerAddAccountCmd`
